make backups folder configurable by the user

+++ lib/mongo-db-utils/config-loader.rb
@@ -1,27 +1,23 @@
 module MongoDbUtils
 
   class ConfigLoader
+    
+    ROOT_FOLDER = "~/.mongo-db-utils"
+    CONFIG_LOCATION = "#{ROOT_FOLDER}/config.yml"
 
-    def self.load( load_path = ".mongo-db-utils/config.yml")
+    def self.load(load_path = CONFIG_LOCATION)
       full_path = self.expand(load_path)
       puts "loading config from #{full_path}"
 
-      if File.exist?(full_path)
+      if File.exist?(full_path) && YAML.load(File.open(full_path))
         config = YAML.load(File.open(full_path))
-        
-        if config == false
-          config = Model::Config.new
-          config.writer = self
-          config.save
-          config
-        else
-          config.writer = self
-          config
-        end
+        config.writer = self
+        config
       else
-        self.initialize_files(full_path)
         config = Model::Config.new
         config.writer = self
+        config.backup_folder = "#{ROOT_FOLDER}/backups"
+        self.initialize_files(full_path)
         File.open( full_path, 'w' ) do |out|
           YAML.dump( config, out )
         end
@@ -29,20 +25,22 @@ module MongoDbUtils
       end
     end
 
-    def self.flush( path = ".mongo-db-utils/config.yml" )
+    def self.flush(path = CONFIG_LOCATION)
       path = self.expand(path)
       puts "removing: #{path}"
       FileUtils.rm(path) if File.exist?(path)
-      puts "flush::success? #{!File.exist?(path)}"
       self.initialize_files(path)
     end
 
-    def self.save(config, path = ".mongo-db-utils/config.yml")
+    def self.save(config, path = CONFIG_LOCATION)
       self._save(config, self.expand(path))
     end
 
     private 
     def self._save(config,full_path)
+      
+      raise "config is nil" if config.nil?
+
       File.open( full_path, 'w' ) do |out|
         YAML.dump( config, out )
       end
